collapse retrieve and backup data into the same activity for legal reasons
